# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventApp API",
        "description": "API REST pour la gestion d'événements collaboratifs - Plateforme de création, gestion et inscription à des événements",
        "version": "1.0.0",
        "contact": {
            "name": "EventApp Support",
            "email": "support@eventapp.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "R\u00e9cup\u00e9rer le profil utilisateur",
                "operationId": "get_api_me",
                "responses": {
                    "200": {
                        "description": "Profil utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "user@example.com"
                                                },
                                                "firstName": {
                                                    "type": "string",
                                                    "example": "Jean"
                                                },
                                                "lastName": {
                                                    "type": "string",
                                                    "example": "Dupont"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifi\u00e9"
                    }
                }
            }
        },
...
        "/api/users/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Mon profil",
                "operationId": "get_api_users_profile",
                "responses": {
                    "200": {
                        "description": "Profil utilisateur"
                    },
                    "401": {
                        "description": "Non authentifi\u00e9"
                    }
                }
            }
        },
        "/api/users/me/events": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Mes \u00e9v\u00e9nements organis\u00e9s",
                "operationId": "get_api_users_my_events",
                "responses": {
                    "200": {
                        "description": "\u00c9v\u00e9nements organis\u00e9s"
                    },
                    "401": {
                        "description": "Non authentifi\u00e9"
                    }
                }
            }
        },
...
        "/api/users/search": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Rechercher des utilisateurs",
                "operationId": "get_api_users_search",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "R\u00e9sultats de recherche"
                    },
                    "403": {
                        "description": "Acc\u00e8s refus\u00e9"
                    }
                }
            }
        },
        "/api/users/stats": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Statistiques des utilisateurs",
                "operationId": "get_api_users_stats",
                "responses": {
                    "200": {
                        "description": "Statistiques"
                    },
                    "403": {
                        "description": "Acc\u00e8s refus\u00e9"
                    }
                }
            }
        }
    },
...
}